+++ consensus/XDPoS/utils/errors.go
@@ -45,6 +45,8 @@ var (
 
 	ErrInvalidCheckpointPenalties = errors.New("invalid penalty list on checkpoint block")
 
+	ErrValidatorsNotLegit = errors.New("Validators does not match what's stored in snapshot minutes its penalty")
+
 	// errInvalidMixDigest is returned if a block's mix digest is non-zero.
 	ErrInvalidMixDigest = errors.New("non-zero mix digest")
 
@@ -80,11 +82,14 @@ var (
 
 	ErrEmptyEpochSwitchValidators = errors.New("empty validators list on epoch switch block")
 
-	ErrInvalidV2Extra               = errors.New("Invalid v2 extra in the block")
-	ErrInvalidQC                    = errors.New("Invalid QC content")
-	ErrInvalidTC                    = errors.New("Invalid TC content")
-	ErrEmptyBlockInfoHash           = errors.New("BlockInfo hash is empty")
-	ErrInvalidFieldInNonEpochSwitch = errors.New("Invalid field exist in a non-epoch swtich block")
+	ErrInvalidV2Extra                = errors.New("Invalid v2 extra in the block")
+	ErrInvalidQC                     = errors.New("Invalid QC content")
+	ErrInvalidTC                     = errors.New("Invalid TC content")
+	ErrEmptyBlockInfoHash            = errors.New("BlockInfo hash is empty")
+	ErrInvalidFieldInNonEpochSwitch  = errors.New("Invalid field exist in a non-epoch swtich block")
+	ErrValidatorNotWithinMasternodes = errors.New("Validaotor address is not in the master node list")
+	ErrCoinbaseAndValidatorMismatch  = errors.New("Validaotor and coinbase address in header does not match")
+	ErrNotItsTurn                    = errors.New("Not validator's turn to mine this block")
 
 	ErrPenaltyListDoesNotMatch = errors.New("Incoming block penalty list does not match")
 	ErrRoundInvalid            = errors.New("Invalid Round, it shall be bigger than QC round")
