+++ consensus/tests/engine_v1_tests/blockchain_race_condition_test.go
@@ -1,4 +1,4 @@
-package tests
+package engine_v1_tests
 
 import (
 	"math/big"
@@ -7,12 +7,13 @@ import (
 	"github.com/XinFinOrg/XDPoSChain/common"
 	"github.com/XinFinOrg/XDPoSChain/core/types"
 	"github.com/XinFinOrg/XDPoSChain/params"
+	"github.com/stretchr/testify/assert"
 )
 
 // Snapshot try to read before blockchain is written
 func TestRaceConditionOnBlockchainReadAndWrite(t *testing.T) {
 
-	blockchain, backend, parentBlock, _ := PrepareXDCTestBlockChain(t, GAP-1, params.TestXDPoSMockChainConfig)
+	blockchain, backend, parentBlock, signer, signFn := PrepareXDCTestBlockChain(t, GAP-1, params.TestXDPoSMockChainConfig)
 
 	state, err := blockchain.State()
 	if err != nil {
@@ -48,11 +49,12 @@ func TestRaceConditionOnBlockchainReadAndWrite(t *testing.T) {
 		Coinbase:   common.HexToAddress(blockCoinbaseA),
 	}
 
-	blockA, err := createBlockFromHeader(blockchain, header, []*types.Transaction{tx, transferTransaction})
+	blockA, err := createBlockFromHeader(blockchain, header, []*types.Transaction{tx, transferTransaction}, signer, signFn, blockchain.Config())
 	if err != nil {
 		t.Fatal(err)
 	}
-	blockchain.InsertBlock(blockA)
+	err = blockchain.InsertBlock(blockA)
+	assert.Nil(t, err)
 	state, err = blockchain.State()
 	if err != nil {
 		t.Fatalf("Failed while trying to get blockchain state")
@@ -93,11 +95,12 @@ func TestRaceConditionOnBlockchainReadAndWrite(t *testing.T) {
 		Difficulty: big.NewInt(2),
 	}
 
-	block450B, err := createBlockFromHeader(blockchain, header, []*types.Transaction{tx, transferTransaction})
+	block450B, err := createBlockFromHeader(blockchain, header, []*types.Transaction{tx, transferTransaction}, signer, signFn, blockchain.Config())
 	if err != nil {
 		t.Fatal(err)
 	}
-	blockchain.InsertBlock(block450B)
+	err = blockchain.InsertBlock(block450B)
+	assert.Nil(t, err)
 	if blockchain.CurrentHeader().Hash() != block450B.Hash() {
 		t.Fatalf("the block with higher difficulty should be current header")
 	}
@@ -131,11 +134,12 @@ func TestRaceConditionOnBlockchainReadAndWrite(t *testing.T) {
 		Coinbase:   common.HexToAddress(blockCoinBase451B),
 		Difficulty: big.NewInt(3),
 	}
-	block451B, err := createBlockFromHeader(blockchain, header, nil)
+	block451B, err := createBlockFromHeader(blockchain, header, nil, signer, signFn, blockchain.Config())
 	if err != nil {
 		t.Fatal(err)
 	}
-	blockchain.InsertBlock(block451B)
+	err = blockchain.InsertBlock(block451B)
+	assert.Nil(t, err)
 
 	signers, err = GetSnapshotSigner(blockchain, block450B.Header())
 	if err != nil {
