+++ consensus/tests/engine_v2_tests/adaptor_test.go
@@ -1,7 +1,6 @@
-package tests
+package engine_v2_tests
 
 import (
-	"fmt"
 	"math/big"
 	"reflect"
 	"testing"
@@ -15,7 +14,7 @@ import (
 )
 
 func TestAdaptorShouldGetAuthorForDifferentConsensusVersion(t *testing.T) {
-	blockchain, backend, currentBlock, _, _, _ := PrepareXDCTestBlockChainForV2Engine(t, 900, params.TestXDPoSMockChainConfig, 0)
+	blockchain, _, currentBlock, signer, signFn, _ := PrepareXDCTestBlockChainForV2Engine(t, 900, params.TestXDPoSMockChainConfig, 0)
 	adaptor := blockchain.Engine().(*XDPoS.XDPoS)
 
 	addressFromAdaptor, errorAdaptor := adaptor.Author(currentBlock.Header())
@@ -32,23 +31,22 @@ func TestAdaptorShouldGetAuthorForDifferentConsensusVersion(t *testing.T) {
 	// Insert one more block to make it above 10, which means now we are on v2 of consensus engine
 	// Insert block 901
 
-	blockCoinBase := fmt.Sprintf("0x111000000000000000000000000000000%03d", 901)
 	merkleRoot := "35999dded35e8db12de7e6c1471eb9670c162eec616ecebbaf4fddd4676fb930"
 	header := &types.Header{
 		Root:       common.HexToHash(merkleRoot),
 		Number:     big.NewInt(int64(901)),
 		ParentHash: currentBlock.Hash(),
-		Coinbase:   common.HexToAddress(blockCoinBase),
+		Coinbase:   signer,
 	}
-	err := generateSignature(backend, adaptor, header)
+
+	header.Extra = generateV2Extra(1, currentBlock, signer, signFn)
+
+	block901, err := createBlockFromHeader(blockchain, header, nil, signer, signFn, blockchain.Config())
 	if err != nil {
 		t.Fatal(err)
 	}
-	block901, err := createBlockFromHeader(blockchain, header, nil)
-	if err != nil {
-		t.Fatal(err)
-	}
-	blockchain.InsertBlock(block901)
+	err = blockchain.InsertBlock(block901)
+	assert.Nil(t, err)
 
 	addressFromAdaptor, errorAdaptor = adaptor.Author(block901.Header())
 	if errorAdaptor != nil {
@@ -181,14 +179,16 @@ func TestAdaptorGetMasternodesV2(t *testing.T) {
 	currentBlock = CreateBlock(blockchain, params.TestXDPoSMockChainConfig, currentBlock, blockNum, 1, blockCoinBase, signer, signFn, nil)
 
 	// block 901 is the first v2 block, and is treated as epoch switch block
-	blockchain.InsertBlock(currentBlock)
+	err := blockchain.InsertBlock(currentBlock)
+	assert.Nil(t, err)
 	masternodes1 := adaptor.GetMasternodes(blockchain, currentBlock.Header())
-	assert.Equal(t, 4, len(masternodes1))
+	assert.Equal(t, 5, len(masternodes1))
 	masternodes1ByNumber := adaptor.GetMasternodesByNumber(blockchain, currentBlock.NumberU64())
 	assert.True(t, reflect.DeepEqual(masternodes1, masternodes1ByNumber), "at block number", blockNum)
 	for blockNum = 902; blockNum < 915; blockNum++ {
 		currentBlock = CreateBlock(blockchain, params.TestXDPoSMockChainConfig, currentBlock, blockNum, int64(blockNum-900), blockCoinBase, signer, signFn, nil)
-		blockchain.InsertBlock(currentBlock)
+		err = blockchain.InsertBlock(currentBlock)
+		assert.Nil(t, err)
 		masternodes2 := adaptor.GetMasternodes(blockchain, currentBlock.Header())
 		assert.True(t, reflect.DeepEqual(masternodes1, masternodes2), "at block number", blockNum)
 		masternodes2ByNumber := adaptor.GetMasternodesByNumber(blockchain, currentBlock.NumberU64())
@@ -210,7 +210,8 @@ func TestGetCurrentEpochSwitchBlock(t *testing.T) {
 	blockNum := 901
 	blockCoinBase := "0x111000000000000000000000000000000123"
 	currentBlock = CreateBlock(blockchain, params.TestXDPoSMockChainConfig, currentBlock, blockNum, 1, blockCoinBase, signer, signFn, nil)
-	blockchain.InsertBlock(currentBlock)
+	err = blockchain.InsertBlock(currentBlock)
+	assert.Nil(t, err)
 	currentCheckpointNumber, epochNum, err = adaptor.GetCurrentEpochSwitchBlock(blockchain, currentBlock.Number())
 	assert.Nil(t, err)
 	assert.Equal(t, uint64(901), currentCheckpointNumber)
@@ -219,7 +220,8 @@ func TestGetCurrentEpochSwitchBlock(t *testing.T) {
 	for blockNum = 902; blockNum < 915; blockNum++ {
 		currentBlock = CreateBlock(blockchain, params.TestXDPoSMockChainConfig, currentBlock, blockNum, int64(blockNum-900), blockCoinBase, signer, signFn, nil)
 
-		blockchain.InsertBlock(currentBlock)
+		err = blockchain.InsertBlock(currentBlock)
+		assert.Nil(t, err)
 		currentCheckpointNumber, epochNum, err := adaptor.GetCurrentEpochSwitchBlock(blockchain, currentBlock.Number())
 		assert.Nil(t, err)
 		assert.Equal(t, uint64(901), currentCheckpointNumber)
@@ -243,13 +245,14 @@ func TestGetParentBlock(t *testing.T) {
 	blockNum := 901
 	blockCoinBase := "0x111000000000000000000000000000000123"
 	block901 := CreateBlock(blockchain, params.TestXDPoSMockChainConfig, block900, blockNum, 1, blockCoinBase, signer, signFn, nil)
-	blockchain.InsertBlock(block901)
+	err = blockchain.InsertBlock(block901)
+	assert.Nil(t, err)
 
 	// let's inject another one, but the highestedQC has not been updated, so it shall still point to 900
 	blockNum = 902
 	block902 := CreateBlock(blockchain, params.TestXDPoSMockChainConfig, block901, blockNum, 1, blockCoinBase, signer, signFn, nil)
-	blockchain.InsertBlock(block902)
-
+	err = blockchain.InsertBlock(block902)
+	assert.Nil(t, err)
 	block = adaptor.FindParentBlockToAssign(blockchain, block902)
 
 	assert.Equal(t, block900.Hash(), block.Hash())
