+++ consensus/tests/engine_v2_tests/authorised_masternode_test.go
@@ -1,76 +1,14 @@
-package tests
+package engine_v2_tests
 
 import (
-	"math/big"
 	"testing"
 	"time"
 
 	"github.com/XinFinOrg/XDPoSChain/common"
 	"github.com/XinFinOrg/XDPoSChain/consensus/XDPoS"
-	"github.com/XinFinOrg/XDPoSChain/core/types"
 	"github.com/XinFinOrg/XDPoSChain/params"
 	"github.com/stretchr/testify/assert"
 )
-
-func TestIsAuthorisedMNForConsensusV1(t *testing.T) {
-	/*
-		V1 consensus engine
-	*/
-	blockchain, _, parentBlock, _ := PrepareXDCTestBlockChain(t, GAP-2, params.TestXDPoSMockChainConfig)
-	// Insert first Block 449
-	t.Logf("Inserting block with propose at 449...")
-	blockCoinbaseA := "0xaaa0000000000000000000000000000000000449"
-	tx, err := voteTX(37117, 0, acc1Addr.String())
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	//Get from block validator error message
-	merkleRoot := "46234e9cd7e85a267f7f0435b15256a794a2f6d65cc98cdbd21dcd10a01d9772"
-	header := &types.Header{
-		Root:       common.HexToHash(merkleRoot),
-		Number:     big.NewInt(int64(449)),
-		ParentHash: parentBlock.Hash(),
-		Coinbase:   common.HexToAddress(blockCoinbaseA),
-	}
-	block449, err := createBlockFromHeader(blockchain, header, []*types.Transaction{tx})
-	blockchain.InsertBlock(block449)
-	if err != nil {
-		t.Fatal(err)
-	}
-	parentBlock = block449
-
-	// At block 449, we should not update signerList. we need to update it till block 450 gap block.
-	// Acc3 is the default account that is on the signerList
-
-	engine := blockchain.Engine().(*XDPoS.XDPoS)
-	isAuthorisedMN := engine.IsAuthorisedAddress(blockchain, block449.Header(), acc3Addr)
-	assert.True(t, isAuthorisedMN)
-
-	isAuthorisedMN = engine.IsAuthorisedAddress(blockchain, block449.Header(), acc1Addr)
-	assert.False(t, isAuthorisedMN)
-
-	// Now, let's mine another block to trigger the GAP block signerList update
-	block450CoinbaseAddress := "0xaaa0000000000000000000000000000000000450"
-	merkleRoot = "46234e9cd7e85a267f7f0435b15256a794a2f6d65cc98cdbd21dcd10a01d9772"
-	header = &types.Header{
-		Root:       common.HexToHash(merkleRoot),
-		Number:     big.NewInt(int64(450)),
-		ParentHash: parentBlock.Hash(),
-		Coinbase:   common.HexToAddress(block450CoinbaseAddress),
-	}
-	block450, err := createBlockFromHeader(blockchain, header, nil)
-	if err != nil {
-		t.Fatal(err)
-	}
-	blockchain.InsertBlock(block450)
-
-	isAuthorisedMN = engine.IsAuthorisedAddress(blockchain, block450.Header(), acc3Addr)
-	assert.False(t, isAuthorisedMN)
-
-	isAuthorisedMN = engine.IsAuthorisedAddress(blockchain, block450.Header(), acc1Addr)
-	assert.True(t, isAuthorisedMN)
-}
 
 func TestIsAuthorisedMNForConsensusV2(t *testing.T) {
 	// we skip test for v1 since it's hard to make a real genesis block
@@ -79,8 +17,8 @@ func TestIsAuthorisedMNForConsensusV2(t *testing.T) {
 	blockNum := 901
 	blockCoinBase := "0x111000000000000000000000000000000123"
 	currentBlock = CreateBlock(blockchain, params.TestXDPoSMockChainConfig, currentBlock, blockNum, 1, blockCoinBase, signer, signFn, nil)
-	blockchain.InsertBlock(currentBlock)
-
+	err := blockchain.InsertBlock(currentBlock)
+	assert.Nil(t, err)
 	// As long as the address is in the master node list, they are all valid
 	isAuthorisedMN := adaptor.IsAuthorisedAddress(blockchain, currentBlock.Header(), common.HexToAddress("xdc0D3ab14BBaD3D99F4203bd7a11aCB94882050E7e"))
 	assert.True(t, isAuthorisedMN)
@@ -100,8 +38,8 @@ func TestIsYourTurnConsensusV2(t *testing.T) {
 	blockNum := 901
 	blockCoinBase := "0x111000000000000000000000000000000123"
 	currentBlock = CreateBlock(blockchain, params.TestXDPoSMockChainConfig, currentBlock, blockNum, 1, blockCoinBase, signer, signFn, nil)
-	blockchain.InsertBlock(currentBlock)
-
+	err := blockchain.InsertBlock(currentBlock)
+	assert.Nil(t, err)
 	// Less then Mine Period
 	isYourTurn, err := adaptor.YourTurn(blockchain, currentBlock.Header(), common.HexToAddress("xdc0D3ab14BBaD3D99F4203bd7a11aCB94882050E7e"))
 	assert.Nil(t, err)
@@ -124,7 +62,8 @@ func TestIsYourTurnConsensusV2(t *testing.T) {
 	// We continue to grow the chain which will increase the round number
 	blockNum = 902
 	currentBlock = CreateBlock(blockchain, params.TestXDPoSMockChainConfig, currentBlock, blockNum, int64(blockNum-900), blockCoinBase, signer, signFn, nil)
-	blockchain.InsertBlock(currentBlock)
+	err = blockchain.InsertBlock(currentBlock)
+	assert.Nil(t, err)
 	time.Sleep(time.Duration(minePeriod) * time.Second)
 
 	adaptor.EngineV2.SetNewRoundFaker(blockchain, 2, false)
